requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
anyio==3.6.2
argcomplete==3.0.8
attrs==23.1.0
betterproto==2.0.0b5
black==23.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
curlify==2.2.1
datamodel-code-generator==0.18.1
db-client @ git+https://github.com/surovp/db_client.git@88e787cd490064b9db6740c3c8a06bd97bfc685b
distconfig3==1.0.1
dm-api-account @ file:///Users/pavelsurov/Documents/openapi-templates/dm_api_account
dm-api-search @ git+https://github.com/surovp/dm_api_search.git@4a819a729f2a0d70bfe01a84a581a9f77ebe448f
dm-api-search-async @ git+https://github.com/surovp/dm_api_search_async.git@d135cd5df420e7eb3a964eef26bb47bc8b18eabd
dnspython==2.3.0
docopt==0.6.2
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
genson==1.2.2
grpcio==1.54.2
grpcio-tools==1.54.2
grpclib==0.4.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
importlib-resources==5.12.0
inflect==5.6.2
iniconfig==2.0.0
isort==5.12.0
jdcal==1.4.1
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-spec==0.1.4
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
mypy-protobuf==3.4.0
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.1
openpyxl==2.4.11
orm-client @ git+https://github.com/surovp/orm_client.git@9fe8e17cce0fbd04f935cd99deb42a8453c66b27
packaging==23.1
pathable==0.4.3
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
prance==0.22.2.22.0
protobuf==4.23.2
psycopg2-binary==2.9.6
pydantic==1.10.7
PyHamcrest==2.0.4
pyrsistent==0.19.3
PySnooper==1.1.1
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
PyYAML==6.0
records==0.5.3
requests==2.28.2
restclient @ file:///Users/pavelsurov/Documents/restclient/dist/restclient-0.0.1.tar.gz#sha256=fb2a5bf7522109da1807e323ccad440166328226c672a2d65d69f24de132e80b
git+https://github.com/surovp/restclient.git
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.46
stringcase==1.2.0
structlog==23.1.0
tablib==3.4.0
toml==0.10.2
tomli==2.0.1
types-protobuf==4.23.0.1
typing_extensions==4.5.0
urllib3==1.26.15
vyper-config==1.1.1
watchdog==3.0.0
